+++ apps/fe/src/app/(main)/data/[id]/page.tsx
@@ -1,0 +1,13 @@
+export default function DataDetailPage({
+  params,
+}: {
+  params: { id: string };
+}) {
+  return (
+    <main className="flex flex-1 flex-col px-40 pt-40 pb-40">
+      {/* TODO: 파일명 + 용량 + 업로드 시간 헤더 */}
+      {/* TODO: 차트 시각화 영역 */}
+      {/* TODO: 채팅 / 삭제 버튼 */}
+    </main>
+  );
+}
