+++ apps/fe/src/app/(main)/data/page.tsx
@@ -1,0 +1,8 @@
+export default function DataPage() {
+  return (
+    <main className="flex flex-1 flex-col px-40 pt-40 pb-40">
+      <h1 className="text-head2 text-gray-900">데이터 소스</h1>
+      {/* TODO: 파일 목록 테이블, 정렬 드롭다운, 새 CSV 파일 업로드 버튼 */}
+    </main>
+  );
+}
