+++ apps/fe/src/app/(main)/history/page.tsx
@@ -1,0 +1,8 @@
+export default function HistoryPage() {
+  return (
+    <main className="flex flex-1 flex-col px-40 pt-40 pb-40">
+      <h1 className="text-head2 text-gray-900">히스토리</h1>
+      {/* TODO: 분석 히스토리 목록 */}
+    </main>
+  );
+}
