+++ apps/fe/src/app/onboarding/page.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+export default function OnboardingPage() {
+  return (
+    <main className="flex min-h-screen items-center justify-center bg-white">
+      {/* TODO: 3단계 온보딩 (도메인 선택 → 업무리스트 → 사용 툴 선택) */}
+      <p className="text-head2 text-gray-900">온보딩 페이지</p>
+    </main>
+  );
+}
